f'n multiTouch is a bitch
kinda works but kinda buggy, time to move on to something else

+++ app/src/main/java/net/cloudhacking/androidgame2/SceneInfo.java
@@ -11,9 +11,9 @@ public interface SceneInfo {
 
     public float[] getProjectionCamera();
 
-    public int getViewportWidth();
+    public float getViewportWidth();
 
-    public int getViewportHeight();
+    public float getViewportHeight();
 
     public void update();
